requirements.txt
python-dotenv==0.11.0
pika==1.1.0
numpy==1.18.2
networkx==2.4
scipy==1.4.1
ortools==7.5.7466
matplotlib==3.1.3
imageio==2.8.0
pygifsicle==1.0.1
Flask==1.1.1
Flask-Login==0.5.0
Flask-Cors==3.0.8
boto3==1.14.33